some more refactoring and adding css/components
