+++ frontend/src/components/Carousel/index.jsx
