--- frontend/src/components/HomePage/HomePage.css
@@ -1,31 +1,0 @@
-.store-navbar-body {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.store-navbar-content {
-  display: flex;
-  border: solid 1px;
-  padding: 0px 15px 0px 15px;
-  margin: 10px;
-}
-
-.store-navbar-links {
-  display: flex;
-
-}
-
-.store-navbar-links > h3 {
-  padding: 0px 10px 0px 10px;
-  font-size: 12px;
-
-}
-
-.homepage-banner-img {
-
-}
-
-.homepage-gutter-block {
-  width: 200px
-}
