+++ frontend/src/components/HomePageGutter/index.jsx
@@ -1,28 +1,10 @@
-import './HomePage.css'
-// eslint-disable-next-line
-import React, { useState } from 'react'
-import Carousel from '../Carousel/Carousel'
+import React, { Component } from 'react'
+import './HomePageGutter.css'
 
-function HomePage() {
-  return (
-    <div className='homepage-wrapper'>
-      <div className="homepage-banner-body">
-        <img src="" alt="" className="homepage-banner-img" />
-        <div className="store-navbar-body">
-          <div className="store-navbar-content">
-            <div className="store-navbar-links">
-              <h3>Your Store</h3>
-              <h3>New & Noteworthy</h3>
-              <h3>Categories</h3>
-              <h3>Points Shop</h3>
-              <h3>News</h3>
-              <h3>Labs</h3>
-            </div>
-          </div>
-        </div>
-      </div>
-      <div className="homepage-body">
-        <div className="homepage-gutter">
+export default class index extends Component {
+  render() {
+    return (
+      <div className="homepage-gutter">
           <div className="homepage-gutter-block">
             <div className="homepage-gutter-header">
               <a href="?"></a>
@@ -70,13 +52,6 @@ function HomePage() {
             </div>
           </div>
         </div>
-        <Carousel/>
-        <div className="homepage-featured-recommended">
-
-        </div>
-      </div>
-    </div>
-  )
+    )
+  }
 }
-
-export default HomePage
